chore: Chatting Response 폴더링 변경
